requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
paho-mqtt==1.6.1
pydantic-settings==2.0.3
pydantic[email]
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
pytest==7.4.0
pytest-html==4.0.0
pytest-cov==4.1.0
pytest-asyncio == 0.21.1
pytest-mock == 3.11.1
pytest-env==1.0.1
httpx==0.25.0
python-dotenv==1.0.0